Move SceneVariableDialog next to other variable dialogs.
